requirements.txt
streamlit
streamlit-lottie
Pillow
python-dotenv
google-generativeai
requests